requirements.txt
python-telegram-bot
textblob